+++ src/support.c
@@ -89,11 +89,11 @@ uint32_t div2freq(uint32_t system_clock, uint16_t div)
 }
 
 /* Builds a buffer of commands + data blocks */
-unsigned char *build_block_buffer(struct mpsse_context *mpsse, uint8_t cmd, unsigned char *data, int size, int *buf_size)
+unsigned char *build_block_buffer(struct mpsse_context *mpsse, uint8_t cmd, unsigned char *data, size_t size, int *buf_size)
 {
 	unsigned char *buf = NULL;
-       	int i = 0, j = 0, k = 0, dsize = 0, num_blocks = 0, total_size = 0, xfer_size = 0;
- 	uint16_t rsize = 0;
+	int i = 0, j = 0, k = 0, dsize = 0, num_blocks = 0, total_size = 0, xfer_size = 0;
+	uint16_t rsize = 0;
 
 	*buf_size = 0;
 
@@ -143,7 +143,7 @@ unsigned char *build_block_buffer(struct mpsse_context *mpsse, uint8_t cmd, unsi
 			{
 				buf[i++] = SET_BITS_LOW;
 				buf[i++] = mpsse->pstart & ~SK;
-				
+
 				/* On receive, we need to ensure that the data out line is set as an input to avoid contention on the bus */
 				if(cmd == mpsse->rx)
 				{
